Suggestions on refactoring home page
