+++ frontend/app/components/header/header.module.scss
@@ -1,0 +1,75 @@
+@import "../../variables.module.scss";
+
+@import "../../mixins/media_query_mixins.scss";
+
+.header {
+  display: flex;
+  align-items: center;
+  gap: 50px;
+  height: 80px;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 9999;
+  background-color: white;
+
+  padding: 0 32px 0 32px;
+  border-bottom: 1px $borderColor solid;
+  .menuLogo {
+    display: none;
+  }
+
+  @include mobile-media-query {
+    height: 60px;
+    gap: 20px;
+
+    .menuLogo {
+      display: block;
+    }
+  }
+
+  input {
+    padding: 10px 10px 10px 50px;
+    width: 100%;
+  }
+
+  .inputGroup {
+    position: relative;
+    width: 60%;
+
+    span {
+      position: absolute;
+      top: 25%;
+      left: 2.5%;
+    }
+
+    @include mobile-media-query {
+      display: none;
+    }
+  }
+
+  button {
+    padding: 10px;
+    width: 10%;
+
+    @include mobile-media-query {
+      width: 20%;
+      margin-left: auto;
+      margin-right: 50px;
+    }
+  }
+}
+
+.rightInactive {
+  display: none;
+
+  @include mobile-media-query {
+    display: block;
+    height: 100vh;
+    width: 100%;
+    position: fixed;
+    z-index: 8000;
+    top: 0;
+    background-color: rgba(0, 0, 0, 0.2);
+  }
+}
